Sting counter handled commmit

+++ app/src/main/java/com/boryans/beenote/model/Sting.kt
@@ -1,5 +1,5 @@
 package com.boryans.beenote.model
 
  data class Sting(
-    var stings: Int = 0
+     var stings: Long = 0
 )
